Changed out the frameworks being used to unzip files; this will allow us to password protect archives if we want.  Switched from CocoaPods to Carthage for easier dependency management.  Added support to the javascript file to handle IBT and CBT courses. Added TrainingType Enum to facilitate. Added Assets.zip as a bundle to reduce the need for the user to get this file. Broke up the function within the "removeFile" function to remove some bugs.  Added icons and color to both the LaunchScreen and the loading view of the WebViewer.

+++ Courseware/Models/CourseCatelog.swift
@@ -7,6 +7,8 @@
 //
 
 import UIKit
+import ZipArchive
+import Zip
 
 class CourseCatelog {
     var allCourses = [Course]()
@@ -41,19 +43,39 @@ class CourseCatelog {
                 }
             }
             
-            guard let urls: [URL] = try? FileManager.default.contentsOfDirectory(at: self.courseScanDirectory, includingPropertiesForKeys: [], options: []) else {
+            guard var urls: [URL] = try? FileManager.default.contentsOfDirectory(at: self.courseScanDirectory, includingPropertiesForKeys: [], options: []) else {
                 return
             }
 
             DispatchQueue.global().async {
                 if (!StaticMethods.doesWebDirectoryExist(self.assetsDirectory)) {
+                    if (urls.filter { $0.lastPathComponent == self.assetsFile }).count == 0 {
+                        if let bundleURL = Bundle.main.url(forResource: "Assets", withExtension: "zip") {
+                            do {
+                                print("Bundle: \(bundleURL)")
+                                let newUrl = self.courseScanDirectory.appendingPathComponent(self.assetsFile)
+                                print("New URL: \(newUrl)")
+                                try FileManager.default.copyItem(at: bundleURL, to: newUrl)
+                                urls.append(newUrl)
+                            } catch {
+                                print("Error copying bundle: \(error)")
+                            }
+                        }
+                    }
+                    
                     guard let assetsPackage = (urls.filter { $0.lastPathComponent == self.assetsFile }).first else {
                         self.hasAssetsFile = false
                         return
                     }
                     
                     self.hasAssetsFile = true
-                    try! FileManager().unzipItem(at: assetsPackage, to: self.courseScanDirectory)
+                
+                    do {
+                        try Zip.unzipFile(assetsPackage, destination: self.courseScanDirectory, overwrite: true, password: nil)
+                    } catch {
+                        print("failed to unzip assets")
+                        print("error: \(error)")
+                    }
                     if StaticMethods.doesWebDirectoryExist(self.assetsDirectory) {
                         print("Worked")
                     } else {
@@ -77,20 +99,24 @@ class CourseCatelog {
         let dispatchGroup = DispatchGroup()
         var shouldReturnJS: Bool = false
         var shouldReturnZip: Bool = false
-        
+        let courseURL = self.allCourses[index].url
+        let trainingType = StaticMethods.isFileCBTorIBT(url: courseURL)
+        guard trainingType != .none else {
+            return nil
+        }
+        let isCBT = trainingType == .cbt
         dispatchGroup.enter()
         DispatchQueue.global().async {
-            shouldReturnJS = StaticMethods.createJavaScriptFile(lessonName: courseName, aircraft: aircraft)
+            shouldReturnJS = StaticMethods.createJavaScriptFile(lessonName: courseName, aircraft: aircraft, isCBT: isCBT)
             dispatchGroup.leave()
         }
 
         dispatchGroup.enter()
         DispatchQueue.global().async {
-            let courseURL = self.allCourses[index].url
             do {
                 let oldCopy = self.mediaDirectory.appendingPathComponent(courseURL.lastPathComponent).absoluteString
                 self.removeFile(fromPath: oldCopy)
-                try FileManager().unzipItem(at: courseURL, to: self.mediaDirectory)
+                try Zip.unzipFile(courseURL, destination: self.mediaDirectory, overwrite: true, password: nil)
                 shouldReturnZip = true
             } catch {
                 print("Error during unzip: \(error)")
@@ -110,8 +136,16 @@ class CourseCatelog {
     }
     
     private func removeFile(fromPath: String) {
-        let passedFile = (URL(string: fromPath)!.lastPathComponent).split(separator: ".").first
-        guard let contents = passedFile else {
+        let passedFileURL = URL(string: fromPath)
+        guard let passedFile = passedFileURL else {
+            return
+        }
+        let toSplit = passedFile.lastPathComponent
+        let split = toSplit.split(separator: ".")
+        guard split.count > 0 else {
+            return
+        }
+        guard let contents = split.first else {
             return
         }
         let contentString = String(contents)
@@ -121,10 +155,11 @@ class CourseCatelog {
             let fileManager = FileManager()
             print(dirToParse.absoluteString)
             let files = try fileManager.contentsOfDirectory(at: dirToParse, includingPropertiesForKeys: nil)
-            for file in files {
-                print("removing item: \(file)")
-//                try fileManager.removeItem(atPath: file)
-                try fileManager.removeItem(at: file)
+            if (files.count > 0) {
+                for file in files {
+                    print("removing item: \(file)")
+                    try fileManager.removeItem(at: file)
+                }
             }
             print("removing the directory")
             try fileManager.removeItem(at: dirToParse)
